Implemented a revised Rules.py file using a more efficient and more scalable vector method to determine where pieces can move.
Old for parts method:
determine if the difference in x and y is a multiple of the direction in which a piece can move.
Then calculate the position of all the cells in that path and make sure there were no pieces.

New vector method:
Each piece has a moveset that are converted to numpy direction vectors. For each possible direction the path is extrapolated outwards and all cells in these paths are checked.
Then ensure the final position is in the list of these positions.

This will make checking for check and checkmate significantly easier and the list of possible positions could be forwarded to the front end to marked on the board